refactor: extract _make_jwt helper to shared tests/helpers.py
Deduplicate the _make_jwt() test helper that was copy-pasted across
test_token_verifier.py and test_oauth_integration.py. Add pythonpath=tests
to pytest.ini so the shared module is importable.
