+++ pytest.ini
@@ -1,6 +1,7 @@
 [pytest]
 asyncio_mode = auto
 testpaths = tests
+pythonpath = tests
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
