+++ tests/helpers.py
@@ -1,0 +1,14 @@
+"""Shared test helpers."""
+
+import base64
+import json
+
+
+def _make_jwt(claims: dict, header: dict | None = None) -> str:
+    """Create a minimal JWT string (unsigned) for testing."""
+    if header is None:
+        header = {"alg": "RS256", "typ": "JWT", "kid": "test-key-id"}
+    h = base64.urlsafe_b64encode(json.dumps(header).encode()).rstrip(b"=").decode()
+    p = base64.urlsafe_b64encode(json.dumps(claims).encode()).rstrip(b"=").decode()
+    s = base64.urlsafe_b64encode(b"fakesig").rstrip(b"=").decode()
+    return f"{h}.{p}.{s}"
