+++ tests/test_oauth_integration.py
@@ -8,23 +8,13 @@ Tests the full integration of:
 - Backward compatibility with legacy service account mode
 """
 
-import base64
-import json
 import os
 import time
 
 import pytest
 from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
-
-def _make_jwt(claims: dict, header: dict | None = None) -> str:
-    """Create a minimal JWT string (unsigned) for testing."""
-    if header is None:
-        header = {"alg": "RS256", "typ": "JWT", "kid": "test-key-id"}
-    h = base64.urlsafe_b64encode(json.dumps(header).encode()).rstrip(b"=").decode()
-    p = base64.urlsafe_b64encode(json.dumps(claims).encode()).rstrip(b"=").decode()
-    s = base64.urlsafe_b64encode(b"fakesig").rstrip(b"=").decode()
-    return f"{h}.{p}.{s}"
+from helpers import _make_jwt
 
 
 def _default_claims(**overrides: object) -> dict:
